requirements.txt
aiohttp==3.8.4
beautifulsoup4==4.12.2
Pyrogram==2.0.106
python-dotenv==1.0.0
TgCrypto==1.2.5